+++ src/components/home/heroSection.js
@@ -3,7 +3,7 @@ import {
   Link,
 } from 'react-router-dom';
 import Styled from 'styled-components';
-import Button from '../button';
+import Button from '../button/index';
 import heatMap from './table.png';
 import searchJson from '../../helper/search.json';
